PlayerStatsOverlayScreen-Funktion in einer eigenen Datei ausgelagert. aus PlayerStatsOverlay.kt und PlayerStatsActivity.kt entfernt
