+++ app/src/main/java/at/aau/serg/sdlapp/ui/PlayerStatsOverlay.kt
@@ -70,26 +70,3 @@ fun StatusIcon(emoji: String, value: String) {
     }
 }
 
-@Composable
-fun PlayerStatsOverlayScreen(
-    playerId: String,
-    viewModel: PlayerViewModel = viewModel()
-) {
-    LaunchedEffect(playerId) {
-        println("PlayerStatsOverlayScreen gestartet mit ID: $playerId")
-        viewModel.loadPlayer(playerId)
-    }
-
-    viewModel.player?.let { player ->
-        println("🎉 Spieler geladen: ${player.id}")
-        PlayerStatsOverlay(player = player)
-    } ?: Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(32.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        println("⌛ Spieler wird noch geladen...")
-        CircularProgressIndicator()
-    }
-}
